Fixes trial fight avatars to use their astral flow only once per fight
